[Store] Split master admin service
