+++ mooncake-store/include/master_admin_service.h
@@ -1,0 +1,76 @@
+#pragma once
+
+#include <atomic>
+#include <cstdint>
+#include <memory>
+#include <mutex>
+#include <optional>
+#include <semaphore>
+#include <string>
+#include <thread>
+
+#include <ylt/coro_http/coro_http_server.hpp>
+
+#include "ha/ha_types.h"
+
+namespace mooncake {
+
+extern const uint64_t kMetricReportIntervalSeconds;
+
+class WrappedMasterService;
+
+class MasterAdminServer {
+   public:
+    MasterAdminServer(uint16_t http_port, bool enable_metric_reporting);
+
+    ~MasterAdminServer();
+
+    bool Start();
+
+    void Stop();
+
+    void SetRuntimeState(ha::MasterRuntimeState state);
+
+    void SetObservedLeader(const std::optional<ha::MasterView>& leader_view);
+
+    void SetServiceDelegate(std::shared_ptr<WrappedMasterService> service);
+
+    void SetServiceAvailable(bool available);
+
+   private:
+    struct RuntimeSnapshot {
+        ha::MasterRuntimeState state = ha::MasterRuntimeState::kStarting;
+        std::optional<ha::MasterView> leader_view;
+        std::shared_ptr<WrappedMasterService> service;
+        bool service_available = false;
+    };
+
+    RuntimeSnapshot SnapshotState() const;
+
+    std::string BuildMetricsText() const;
+
+    std::string BuildMetricsSummaryText() const;
+
+    std::string BuildHealthJson() const;
+
+    std::string BuildLeaderJson() const;
+
+    std::shared_ptr<WrappedMasterService> GetActiveService() const;
+
+    void InitHttpServer();
+
+    uint16_t http_port_;
+    bool enable_metric_reporting_ = false;
+    coro_http::coro_http_server http_server_;
+    std::thread metric_report_thread_;
+    std::atomic<bool> metric_report_running_{false};
+    std::binary_semaphore metric_report_stop_sem_{0};
+    std::atomic<bool> started_{false};
+    mutable std::mutex state_mutex_;
+    ha::MasterRuntimeState state_{ha::MasterRuntimeState::kStarting};
+    std::optional<ha::MasterView> leader_view_;
+    std::shared_ptr<WrappedMasterService> service_;
+    bool service_available_ = false;
+};
+
+}  // namespace mooncake
